pyproject.toml
[project]
name = "teams-chatbot"
version = "0.1.0"
description = "A Microsoft Teams chatbot powered by LangGraph and LangChain"
authors = [
    {name = "ASMA", email = "asma.liaquat22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "botbuilder-core >=4.17.0,<5.0.0",
    "botbuilder-schema >=4.17.0,<5.0.0",
    "botbuilder-integration-aiohttp >=4.17.0,<5.0.0",
    "fastapi >=0.95.0",
    "uvicorn >=0.22.0",
    "langchain >=0.0.300",
    "langgraph >=0.1.0",
    "pydantic >=2.0",
    "loguru >=0.7.0",
    "httpx >=0.24.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.20.3"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
